% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{haversine_vec}
\alias{haversine_vec}
\title{Vectorized Haversine Distance (Paired)
Calculate distances between corresponding pairs of points.}
\usage{
haversine_vec(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{Vector of starting latitudes}

\item{lon1}{Vector of starting longitudes}

\item{lat2}{Vector of ending latitudes}

\item{lon2}{Vector of ending longitudes}
}
\value{
Vector of distances in meters
}
\description{
Vectorized Haversine Distance (Paired)
Calculate distances between corresponding pairs of points.
}
\examples{
haversine_vec(c(52.09, 52.10), c(5.12, 5.13),
              c(52.10, 52.11), c(5.13, 5.14))
}
