% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loc_data_2019}
\alias{loc_data_2019}
\title{Timestamped Location Data}
\format{
\subsection{\code{loc_data_2019}}{

A data frame with 21,911 rows and 3 columns:
\describe{
\item{latitude}{unprojected latitude coordinate using WGS84 ellipsoid}
\item{longitude}{unprojected longitude coordinate using WGS84 ellipsoid}
\item{timestamp}{POSIXct timestamp with date and time using fractional seconds}
...
}
}
}
\source{
Personal recorded location history
}
\usage{
loc_data_2019
}
\description{
Real data from November 2019 extracted from Google Location History files captured with an Android smartphone. Contains two weeks of human movement behavior of a single person occurring in the Netherlands. Modes include biking, walking, bus and train.
}
\keyword{datasets}
