% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergingCycleFunctions.R
\name{mergeStopsAndCleanTracks}
\alias{mergeStopsAndCleanTracks}
\title{Merge Stops and Clean Tracks}
\usage{
mergeStopsAndCleanTracks(events, thetaD, small_track_action = "merge", ...)
}
\arguments{
\item{events}{data.table from \code{returnStateEvents}}

\item{thetaD}{how many meters away may stops be and still be merged}

\item{small_track_action}{one of "merge" or "exclude" for short tracks}

\item{...}{additional optional arguments passed to moveMerger including max_locs, max_time and max_dist}
}
\value{
update events data.table by reference
}
\description{
Calls \code(stopMerger) and then \code{moveMerger}
}
