% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergingCycleFunctions.R
\name{mergingCycle}
\alias{mergingCycle}
\title{Merging Cycle}
\usage{
mergingCycle(res, max_merges = Inf, thetaD = 200, small_track_action = "merge")
}
\arguments{
\item{res}{Results data.table from \code{stopFinder}}

\item{max_merges}{failsafe to prevent infinite loop}

\item{thetaD}{how many meters away may stops be and still be merged}

\item{small_track_action}{one of "merge" or "exclude" for short tracks}
}
\value{
Modifies res data.table by reference
}
\description{
Runs the stop and merging cycle until no changes are seen or until the max
number of merges are met.
}
