% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopdetection-package.R
\docType{package}
\name{stopdetection-package}
\alias{stopdetection}
\alias{stopdetection-package}
\title{stopdetection: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: McCool Danielle \email{d.m.mccool@uu.nl} (\href{https://orcid.org/0000-0002-7055-7539}{ORCID})

}
\keyword{internal}
