% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopFinder.R
\name{stopFinder}
\alias{stopFinder}
\title{Find an initial set of stops given timestamped locations}
\usage{
stopFinder(traj, thetaD, thetaT)
}
\arguments{
\item{traj}{An ordered data.table with columns named timestamp, longitude and latitude}

\item{thetaD}{The distance parameter, represents a radius in meters for establishing how much area a stop can encompass.}

\item{thetaT}{The time parameter, representing the length of time that must be spent within the stop area before being considered a stop.}
}
\value{
traj is modified by reference to include a column stop_id, which is NA for locations not belonging to a stop, and equal to the row number initiating the stop it belongs to otherwise.
}
\description{
Modifies by reference a data.table of trajectories, which are clustered
spatiotemporally based on a user-provided distance radius parameter and time
parameter. Points are evaluated sequentially to determine whether they meet
the criteria for being a stop (at least \code{thetaT} time spent within
\code{thetaD} distance of the initiating location). Points must therefore
have a timestamp, longitude and latitude column.
}
\details{
This function has been optimized for simulation studies where it will be
called repeatedly. Because of this, all error-handling is done prior to this
step. If calling this function directly, the user must ensure that the data
are ordered based on the timestamp, and that the columns names are correct.
}
\examples{
# Set up data

}
