% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopMerger.R
\name{stopMerger}
\alias{stopMerger}
\title{Stop Merger}
\usage{
stopMerger(events, thetaD)
}
\arguments{
\item{events}{data.table of events from \code{\link{returnStateEvents}}}

\item{thetaD}{maximum distance for merging subsequent stops}
}
\value{
modifies events data.table by reference, changing new_stop_id and
new_state
}
\description{
Given the events data.table containing the spatiotemporally clustered stop/
move states, merges stops separated by less than \code{thetaD} meters.
Modifies events by reference.
}
