% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveMerger.R
\name{moveMerger}
\alias{moveMerger}
\title{Move Merger}
\usage{
moveMerger(events, max_locs = 1, max_time = 600)
}
\arguments{
\item{events}{data.table of events generated by \code{returnStateEvents}}

\item{max_locs}{Maximum number of locations for a track to be mergeable}

\item{max_time}{Maximum time elapsed for a track to be mergeable}
}
\value{
Modifies events by reference
}
\description{
Discards move events (tracks) between stop events that should be discarded
during merging. This is based on the researcher-decided total number of
allowable locations that the discarded track can consist of, as well as a
maximum total time length that may elapse. Tracks are merged into the
preceding stop, which is an arbitrary decision. Future versions of this
should consider either discarding the measurement or assigning it to the
closest stop.
}
