% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveMerger.R
\name{moveMerger}
\alias{moveMerger}
\title{Move Merger}
\usage{
moveMerger(
  events,
  small_track_action = "merge",
  max_locs = 1,
  max_time = 600,
  max_dist = 100
)
}
\arguments{
\item{events}{data.table of events generated by \code{returnStateEvents}}

\item{max_locs}{Maximum number of locations for a track to be mergeable. Set to Inf to not consider.}

\item{max_time}{Maximum time elapsed (seconds) for a track to be mergeable. Set to Inf to not consider.}

\item{max_dist}{Maximum distance (meters) traveled while on track to be mergeable. Set to Inf to not consider.}

\item{action}{One of "merge" or "exclude" for specifying the method of
handling mergeable tracks}
}
\value{
Modifies events data.table by reference
}
\description{
Discards move events (tracks) between stop events that should be discarded
during merging. This is based on the researcher-decided total number of
allowable locations that the discarded track can consist of, as well as a
maximum total time length that may elapse. Tracks are merged into the
preceding stop or excluded. Future versions of this should consider assigning
to the closest stop for action = merge
}
